Roles/Permissions bug fixes
This updates the source code to fix some bug fixes. It focuses on a bug where a new role would be created and the role would be given to a user. If the new role got any permissions, the users who are part of the role will not actually get those permissions and will remain a guest.